Move CompletableFutureCancellationHandler to pulsar-common util
